+++ src/components/Map.tsx
@@ -1,218 +1,359 @@
 import { getDistance } from "geolib";
-import { divIcon, LatLngExpression } from "leaflet";
-import React, { useMemo, useState } from "react";
-import { renderToStaticMarkup } from "react-dom/server";
-import {
-  MapContainer,
-  Marker,
-  Polyline,
-  TileLayer,
-  useMapEvent,
-} from "react-leaflet";
-import { Syria } from "../dcs/maps/Syria";
-import { useKeyPress } from "../hooks/useKeyPress";
+import * as maptalks from "maptalks";
+import React, {
+  MutableRefObject,
+  useEffect,
+  useMemo,
+  useRef,
+  useState,
+} from "react";
+import { DCSMap } from "../dcs/maps/DCSMap";
 import { serverStore } from "../stores/ServerStore";
-import { getBearing, getCardinal } from "../util";
-import { MapEntity } from "./MapEntity";
+import { estimatedSpeed, trackStore } from "../stores/TrackStore";
+import { computeBRAA, getBearing, getCardinal } from "../util";
+import { MapTrackedEntityInner } from "./MapEntity";
 
-function MapObjects() {
-  const entities = serverStore((state) =>
-    state.entities.valueSeq().filter((k) =>
-      (!k.types.includes("Bullseye") || k.coalition !== "Allies") &&
-      !k.types.includes("Parachutist")
-    )
+function Marker(
+  { map, latitude, longitude, children, zoom }: {
+    map: maptalks.Map;
+    latitude: number;
+    longitude: number;
+    children?: React.ReactNode;
+    zoom: number;
+  },
+) {
+  const targetCoord = new maptalks.Coordinate(
+    longitude,
+    latitude,
   );
-  const [activeObjectId, setActiveObjectId] = useState<number | null>(null);
-  const [scale, setScale] = useState<number>(124);
-  const [hideLabels, setHideLabels] = useState<boolean>(false);
-  const zoomEvent = useMapEvent("zoomend", () => {
-    const y = zoomEvent.getSize().y;
-    const x = zoomEvent.getSize().x;
-    var maxMeters = zoomEvent.containerPointToLatLng([0, y]).distanceTo(
-      zoomEvent.containerPointToLatLng([x, y]),
-    );
-    setScale(maxMeters / x);
-  });
-
-  const [braaStartPos, setBraaStartPos] = useState<
-    number | [number, number] | null
-  >(
-    null,
-  );
-  const [cursorPos, setCursorPos] = useState<number | [number, number] | null>(
-    null,
-  );
-  const isSnapDown = useKeyPress("s");
-
-  useMapEvent("contextmenu", (e) => {});
-
-  useMapEvent("keyup", (e) => {
-    if (e.originalEvent.key === "l") {
-      setHideLabels(!hideLabels);
-    }
-  });
-
-  useMapEvent("mousemove", (e) => {
-    let snappedObject = null;
-    if (isSnapDown) {
-      snappedObject = entities.map((
-        obj,
-      ) =>
-        [
-          obj.id,
-          getDistance([obj.latitude, obj.longitude], [
-            e.latlng.lat,
-            e.latlng.lng,
-          ]),
-        ] as [number, number]
-      ).sort((a, b) => a[1] - b[1]).first();
-    }
-
-    if (snappedObject) {
-      setCursorPos(
-        snappedObject[0],
-      );
-    } else {
-      setCursorPos([e.latlng.lat, e.latlng.lng]);
-    }
-  });
-
-  const mouseDownEvent = useMapEvent("mousedown", (e) => {
-    if (e.originalEvent.button === 2) {
-      if (isSnapDown) {
-        const snappedObject = entities.sort((
-          a,
-          b,
-        ) =>
-          getDistance([a.latitude, a.longitude], [
-            e.latlng.lat,
-            e.latlng.lng,
-          ]) - getDistance([b.latitude, b.longitude], [
-            e.latlng.lat,
-            e.latlng.lng,
-          ])
-        ).first();
-        if (snappedObject) {
-          setBraaStartPos(snappedObject.id);
-        }
-      } else {
-        setBraaStartPos([e.latlng.lat, e.latlng.lng]);
-      }
-      mouseUpEvent.dragging.disable();
-    }
-  });
-
-  const braaObj = typeof braaStartPos === "number" &&
-    serverStore.getState().entities.get(braaStartPos);
-  let braaPos: [number, number] | undefined = undefined;
-  if (typeof braaStartPos === "number" && braaObj) {
-    braaPos = [braaObj.latitude, braaObj.longitude];
-  } else if (Array.isArray(braaStartPos)) {
-    braaPos = braaStartPos;
-  }
-
-  const cursorObj = typeof cursorPos === "number" &&
-    serverStore.getState().entities.get(cursorPos);
-  let ourCursorPos: [number, number] | undefined = undefined;
-  if (typeof cursorPos === "number" && cursorObj) {
-    ourCursorPos = [cursorObj.latitude, cursorObj.longitude];
-  } else if (Array.isArray(cursorPos)) {
-    ourCursorPos = cursorPos;
-  }
-
-  const mouseUpEvent = useMapEvent("mouseup", (e) => {
-    if (e.originalEvent.button === 2) {
-      mouseDownEvent.dragging.enable();
-    }
-    if (braaStartPos) {
-      setBraaStartPos(null);
-    }
-  });
-
-  const icon = useMemo(() => {
-    if (!braaPos || !ourCursorPos) {
-      return null;
-    }
-
-    let bearing = Math.round(getBearing(braaPos, ourCursorPos)) +
-      Syria.magDec;
-    if (bearing > 360) {
-      bearing = bearing - 360;
-    } else if (bearing < 0) {
-      bearing = bearing + 360;
-    }
-
-    return divIcon({
-      html: renderToStaticMarkup(
-        <div
-          className="absolute text-indigo-300 ml-10 text-xl whitespace-nowrap bg-gray-600 p-2"
-        >
-          {bearing}
-          {getCardinal(bearing)} / {Math.floor(
-            getDistance(braaPos, ourCursorPos) * 0.00053995680345572,
-          )}NM
-        </div>,
-      ),
-      className: "",
-    });
-  }, [braaStartPos, cursorPos, entities]);
-
+  const targetPosition = map.coordinateToContainerPoint(targetCoord).round();
   return (
-    <>
-      {entities.map((obj) => (
-        <MapEntity
-          key={obj.id}
-          entity={obj}
-          active={obj.id === activeObjectId}
-          setActive={() =>
-            activeObjectId === obj.id
-              ? setActiveObjectId(null)
-              : setActiveObjectId(obj.id)}
-          scale={scale}
-          hideLabel={hideLabels}
-        />
-      ))}
-      {braaPos && ourCursorPos && (
-        <>
-          <Polyline
-            positions={[
-              ourCursorPos,
-              braaPos,
-            ]}
-            pathOptions={{
-              weight: 2,
-              color: "yellow",
-            }}
-          />
-          {icon && (
-            <Marker
-              position={ourCursorPos}
-              icon={icon}
-              zIndexOffset={30}
-            />
-          )}
-        </>
-      )}
-    </>
+    <div
+      onContextMenu={(e) => {
+        e.preventDefault();
+      }}
+      style={{
+        position: "absolute",
+        top: targetPosition.y,
+        left: targetPosition.x,
+        pointerEvents: "none",
+      }}
+    >
+      {children}
+    </div>
   );
 }
 
-export default function Map(): JSX.Element {
+export function Map({ dcsMap }: { dcsMap: DCSMap }) {
+  const mapContainer: MutableRefObject<HTMLDivElement | null> = useRef(null);
+  const map: MutableRefObject<maptalks.Map | null> = useRef(null);
+  const radarTracks = trackStore((state) => state.tracks.entrySeq().toArray());
+  const [zoom, setZoom] = useState<number>(8);
+
+  const [drawBraaStart, setDrawBraaStart] = useState<
+    [number, number] | null
+  >(null);
+  const [cursorPos, setCursorPos] = useState<[number, number] | null>(
+    null,
+  );
+
+  useEffect(() => {
+    if (!mapContainer.current || map.current !== null) {
+      return;
+    }
+
+    var braaLine = new maptalks.LineString([], {
+      id: "braa-line",
+      arrowStyle: null,
+      visible: false,
+      editable: false,
+      cursor: null,
+      shadowBlur: 0,
+      shadowColor: "black",
+      draggable: false,
+      dragShadow: false,
+      drawOnAxis: null,
+      symbol: {
+        "lineColor": "yellow",
+        "lineWidth": 2,
+      },
+    });
+
+    var point = new maptalks.Marker(
+      [0, 0],
+      {
+        id: "braa-text",
+        visible: false,
+        editable: false,
+        shadowBlur: 0,
+        shadowColor: "black",
+        draggable: false,
+        dragShadow: false,
+        drawOnAxis: null,
+        symbol: {},
+      },
+    );
+
+    map.current = new maptalks.Map(mapContainer.current, {
+      panAnimation: false,
+      dragRotate: false,
+      dragPitch: false,
+      touchZoom: false,
+      doubleClickZoom: false,
+      center: [dcsMap.center[1], dcsMap.center[0]],
+      zoom: 8,
+      seamlessZoom: true,
+      baseLayer: new maptalks.TileLayer("base", {
+        urlTemplate:
+          "https://cartodb-basemaps-{s}.global.ssl.fastly.net/dark_all/{z}/{x}/{y}.png",
+        subdomains: ["a", "b", "c"],
+      }),
+      layers: [
+        new maptalks.VectorLayer("braa", [braaLine, point]),
+        new maptalks.VectorLayer("tracks", []),
+        new maptalks.VectorLayer("trails", []),
+      ],
+    } as any);
+
+    map.current.on("contextmenu", (e) => {
+    });
+
+    map.current.on("zooming", (e) => {
+      setZoom(map.current!.getZoom());
+    });
+
+    map.current.on("mousemove", (e) => {
+      setCursorPos([e.coordinate.y, e.coordinate.x]);
+    });
+
+    map.current.on("mousedown", (e) => {
+      if (e.domEvent.button === 2) {
+        setDrawBraaStart(
+          [e.coordinate.y, e.coordinate.x],
+        );
+      }
+    });
+
+    map.current.on("mouseup", (e) => {
+      if (e.domEvent.button === 2) {
+        setDrawBraaStart(null);
+      }
+    });
+  }, [mapContainer, map]);
+
+  useEffect(() => {
+    if (!map.current) return;
+    const braaLayer = map.current.getLayer("braa") as maptalks.VectorLayer;
+    const line = braaLayer.getGeometryById("braa-line") as maptalks.LineString;
+    const text = braaLayer.getGeometryById("braa-text") as maptalks.Marker;
+
+    if (drawBraaStart && cursorPos) {
+      let start = drawBraaStart;
+      let end = cursorPos;
+      if (typeof start !== "number" && typeof end !== "number") {
+        let bearing = Math.round(getBearing(start, end)) +
+          dcsMap.magDec;
+        if (bearing > 360) {
+          bearing = bearing - 360;
+        } else if (bearing < 0) {
+          bearing = bearing + 360;
+        }
+        line.setCoordinates([
+          [start[1], start[0]],
+          [end[1], end[0]],
+        ]);
+        text.setCoordinates([end[1], end[0]]);
+        text.setSymbol({
+          "textPlacement": "line",
+          "textFaceName": '"microsoft yahei"',
+          "textName": `${bearing}${getCardinal(bearing)} / ${
+            Math.floor(
+              getDistance(start, end) * 0.00053995680345572,
+            )
+          }NM`,
+          "textFill": "#A5B4FC",
+          "textHorizontalAlignment": "right",
+          "textSize": 14,
+          "textDx": 15,
+        });
+        text.show();
+        line.show();
+      }
+    } else {
+      text.hide();
+      line.hide();
+    }
+  }, [drawBraaStart, cursorPos]);
+
+  useEffect(() => {
+    if (!map.current) return;
+    const entities = serverStore.getState().entities;
+
+    const layer = map.current.getLayer("tracks") as maptalks.VectorLayer;
+    const trailLayer = map.current.getLayer("trails") as maptalks.VectorLayer;
+    for (const geo of layer.getGeometries()) {
+      if (!entities.has(geo.id as number)) {
+        geo.remove();
+      }
+    }
+
+    for (const geo of trailLayer.getGeometries()) {
+      const geoA: any = geo;
+      if (!geoA._id) continue;
+      const [geoId, _] = (geoA._id as string).split("-");
+      if (!entities.has(parseInt(geoId))) {
+        geo.remove();
+      }
+    }
+
+    for (const [entityId, track] of radarTracks) {
+      const trackVisible = estimatedSpeed(track) >= 15;
+      const entity = entities.get(entityId);
+      if (!entity) {
+        continue;
+      }
+
+      let index = 0;
+      for (const trackPoint of track.slice(1)) {
+        const trackPointGeo = trailLayer.getGeometryById(
+          `${entityId}-${index}`,
+        ) as maptalks.Marker;
+        if (!trackPointGeo) {
+          trailLayer.addGeometry(
+            new maptalks.Marker([
+              trackPoint.position[1],
+              trackPoint.position[0],
+            ], {
+              id: `${entityId}-${index}`,
+              visible: trackVisible,
+              editable: false,
+              cursor: "pointer",
+              shadowBlur: 0,
+              draggable: false,
+              dragShadow: false,
+              drawOnAxis: null,
+              symbol: {},
+            }),
+          );
+        } else {
+          if (trackVisible) {
+            trackPointGeo.show();
+          } else {
+            trackPointGeo.hide();
+          }
+          trackPointGeo.setCoordinates([
+            trackPoint.position[1],
+            trackPoint.position[0],
+          ]);
+          trackPointGeo.setSymbol(
+            {
+              "markerType": "square",
+              "markerFill": !trackVisible || track.length < 5
+                ? "white"
+                : entity.coalition !== "Allies"
+                ? "#17c2f6"
+                : "#ff8080",
+              "markerLineOpacity": 0,
+              "markerLineDasharray": [],
+              "markerWidth": 5,
+              "markerHeight": 5,
+              "markerDx": 0,
+              "markerDy": 0,
+              "markerFillOpacity": (100 - (index * 10)) / 100,
+            },
+          );
+        }
+
+        index++;
+      }
+
+      const speed = track && estimatedSpeed(track);
+      const dirArrowEnd = speed && speed >= 15 && track && track.length >= 5 &&
+        computeBRAA(
+          track[0].position[0],
+          track[0].position[1],
+          track[0].heading,
+          // knots -> meters per second -> 30 seconds
+          ((speed * 0.514444)) * 30,
+        );
+      const geo = layer.getGeometryById(
+        entityId,
+      ) as (maptalks.LineString | null);
+
+      if (dirArrowEnd) {
+        if (!geo) {
+          layer.addGeometry(
+            new maptalks.LineString([
+              [track[0].position[1], track[0].position[0]],
+              [dirArrowEnd[1], dirArrowEnd[0]],
+            ], {
+              id: entityId,
+              arrowStyle: "classic",
+              arrowPlacement: "vertex",
+              symbol: {
+                "lineColor": entity.coalition !== "Allies"
+                  ? "#17c2f6"
+                  : "#ff8080",
+                "lineWidth": 1,
+              },
+            }),
+          );
+        } else {
+          geo.setCoordinates([
+            [track[0].position[1], track[0].position[0]],
+            [dirArrowEnd[1], dirArrowEnd[0]],
+          ]);
+        }
+        geo?.show();
+      } else {
+        geo?.hide();
+        // TODO: idk
+      }
+    }
+  }, [radarTracks]);
+
+  const radarTrackEls = useMemo(
+    () =>
+      map.current && radarTracks.map(([entityId, track]) => {
+        const entity = serverStore.getState().entities.get(entityId);
+        if (!entity) {
+          return <></>;
+        }
+
+        return (
+          <Marker
+            latitude={entity.latitude}
+            longitude={entity.longitude}
+            zoom={zoom}
+            map={map.current!}
+            key={entityId}
+          >
+            <MapTrackedEntityInner
+              entity={entity}
+              active={false}
+              track={track}
+              hideLabel={false}
+            />
+          </Marker>
+        );
+      }),
+    [radarTracks],
+  );
+
   return (
-    <MapContainer
-      doubleClickZoom={false}
-      center={Syria.center as LatLngExpression}
-      zoom={9}
-      minZoom={8}
-      maxZoom={12}
-      scrollWheelZoom={true}
-      className="h-full w-full relative"
-    >
-      <TileLayer
-        attribution='&copy; <a href="https://stadiamaps.com/">Stadia Maps</a>, &copy; <a href="https://openmaptiles.org/">OpenMapTiles</a> &copy; <a href="http://openstreetmap.org">OpenStreetMap</a> contributors'
-        url="https://tiles.stadiamaps.com/tiles/alidade_smooth_dark/{z}/{x}/{y}{r}.png"
-      />
-      <MapObjects />
-    </MapContainer>
+    <div>
+      <div
+        style={{
+          display: "block",
+          position: "absolute",
+          top: 0,
+          bottom: 0,
+          width: "100%",
+        }}
+        ref={mapContainer}
+      >
+      </div>
+      {radarTrackEls}
+    </div>
   );
 }
